backend/requirements.txt
fastapi
fastapi-users[sqlalchemy]
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-multipart
pytest
pytest-asyncio
aiosqlite
asyncpg
python-dotenv
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
email-validator
phonenumbers
redis
psutil
